Add equality operations for Expressions
Can now check if two Expressions are equal by checking if all of their
properties are equal.

For Multiplication and Addition, commutation is enabled
(i.e. they're equal if a.left == b.right && a.right == b.left).

For Numbers, comparison between a Number expression and a raw number
(i.e. int or float) is allowed.
For example, Number(2) == 2 == 2.0